% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_issue_authors.R
\name{collect_issue_authors}
\alias{collect_issue_authors}
\title{Collect authors of issues in a repo}
\usage{
collect_issue_authors(owner, repo)
}
\arguments{
\item{owner}{owner of the repo, string}

\item{repo}{repo name, string}
}
\value{
a character vector of all logins of authors of issues
}
\description{
Collect authors of issues in a repo
}
\examples{
collect_issue_authors("ropensci", "taxize")
}
