% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_collaborators.R
\name{get_collaborators}
\alias{get_collaborators}
\title{Get collaborators of a repository}
\usage{
get_collaborators(owner, repo)
}
\arguments{
\item{owner}{owner of the repo, string}

\item{repo}{repo name, string}
}
\value{
a data.frame of collaborators name, login and rights
}
\description{
Get collaborators of a repository
}
\examples{
\dontrun{
get_collaborators("ropensci", "ropenaq")
}
}
