% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contents.R
\name{get_contents}
\alias{get_contents}
\title{Get contents of a repo}
\usage{
get_contents(owner, repo)
}
\arguments{
\item{owner}{owner of the repo, string}

\item{repo}{repo name, string}
}
\value{
a character vector of the top-level contents of the repo.
}
\description{
Get contents of a repo
}
\details{
thanks to https://gist.github.com/jonathansick/8bbe88a85addaeeea4e7fe9ef15b016b
}
\examples{
get_contents("maelle", "convertagd")
}
