% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contributions.R
\name{get_contributions}
\alias{get_contributions}
\title{Get contributions by a given account}
\usage{
get_contributions(privacy = "PUBLIC")
}
\arguments{
\item{privacy}{string, one of "PUBLIC", "PRIVATE", or "null" for both.}
}
\value{
tibble with name (owner/repo), description, primary language, stargazer count, is_fork, is_privatem, and is_created.
}
\description{
Get contributions by a given account
}
\examples{
\dontrun{
cb <- get_contributions()
}
}
