% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_description.R
\name{get_description}
\alias{get_description}
\title{Get description of the package repo}
\usage{
get_description(owner, repo, branch = "master")
}
\arguments{
\item{branch}{master by default}
}
\value{
a tibble of basic information from the DESCRIPTION
}
\description{
Get description of the package repo
}
\details{
https://platform.github.community/t/query-repo-contents-with-graphql-api/1896
}
\examples{
library("magrittr")
get_description(owner = "ropensci", repo = "ropenaq")
}
