% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file.R
\name{get_file}
\alias{get_file}
\title{Get contents of a file}
\usage{
get_file(owner, repo, branch, path)
}
\arguments{
\item{owner}{owner of the repo, string}

\item{repo}{repo name, string}

\item{branch}{master by default}

\item{path}{path to the file}
}
\value{
nothing interesting yet
}
\description{
Get contents of a file
}
\details{
https://platform.github.community/t/query-repo-contents-with-graphql-api/1896
}
\examples{
library("magrittr")
pof <- get_file(owner = "ropensci", repo = "roregistry",
         branch = "master", path = "registry.json") \%>\%
         jsonlite::toJSON()
}
