% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_issues_labels_state.R
\name{get_issue_labels_state}
\alias{get_issue_labels_state}
\title{Table of issues info}
\usage{
get_issue_labels_state(owner, repo)
}
\arguments{
\item{owner}{owner of the repo, string}

\item{repo}{repo name, string}
}
\value{
a tibble of issues info, currently stupidly labels are just concatenated.
}
\description{
Table of issues info
}
\examples{
get_issue_labels_state("ropensci", "taxize")
}
