% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_issue_thread.R
\name{get_issue_thread}
\alias{get_issue_thread}
\title{Get a whole issue thread}
\usage{
get_issue_thread(owner, repo, issue_id)
}
\arguments{
\item{owner}{owner of the repo, string}

\item{repo}{repo name, string}

\item{issue_id}{issue ID}
}
\value{
A data.frame with one row per comment and
 title (of the issue), author, author_association,
assignee, created_at (for each comment), closed_at (for the whole issue),
body of each comment, comment_url, one column per label, issue.
}
\description{
Get a whole issue thread
}
\examples{
get_issue_thread("ropensci", "onboarding", 175)
}
