% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_issues_thumbs.R
\name{get_issues_thumbs}
\alias{get_issues_thumbs}
\title{get table of issues ordered by number of thumbs up}
\usage{
get_issues_thumbs(owner, repo)
}
\arguments{
\item{owner}{owner of the repo, string}

\item{repo}{repo name, string}
}
\value{
a tibble
}
\description{
get table of issues ordered by number of thumbs up
}
\examples{
get_issues_thumbs("tidyverse", "dplyr")
}
