% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_master_committers.R
\name{get_master_committers}
\alias{get_master_committers}
\title{Get committers to master}
\usage{
get_master_committers(repo, owner)
}
\arguments{
\item{repo}{repo name, string}

\item{owner}{owner of the repo, string}
}
\value{
data.frame of login and name
}
\description{
Get committers to master
}
