% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contributions.R
\name{get_repos_contributed}
\alias{get_repos_contributed}
\title{Get repos contributed to by a given account}
\usage{
get_repos_contributed(user, privacy = "PUBLIC")
}
\arguments{
\item{user}{username}

\item{privacy}{string, one of "PUBLIC", "PRIVATE", or "null" for both.}
}
\value{
tibble with name (owner/repo), description, primary language, stargazer count, is_fork, is_private, and is_created.
}
\description{
Get repos contributed to by a given account
}
\examples{
\dontrun{
cb <- get_repos_contributed()
}
}
