% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_teams.R
\name{get_teams}
\alias{get_teams}
\title{Get teams of an organisation}
\usage{
get_teams(owner)
}
\arguments{
\item{owner}{string}
}
\value{
a data.frame of teams info, repositories names
and member logins as nested data.frames.
}
\description{
Get teams of an organisation
}
\examples{
\dontrun{
get_teams("ropensci")
}
}
