% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_travis_config.R
\name{get_travis_config}
\alias{get_travis_config}
\title{Get Travis config file of the package repo}
\usage{
get_travis_config(owner, repo, branch = "master")
}
\arguments{
\item{branch}{master by default}
}
\value{
a tibble of basic information from the Travis config file
}
\description{
Get Travis config file of the package repo
}
\details{
https://platform.github.community/t/query-repo-contents-with-graphql-api/1896
}
\examples{
library("magrittr")
get_travis_config(owner = "ropensci", repo = "ropenaq")
}
