% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_package.R
\name{is_package_repo}
\alias{is_package_repo}
\title{Is this repo a package}
\usage{
is_package_repo(owner, repo)
}
\arguments{
\item{owner}{owner of the repo, string}

\item{repo}{repo name, string}
}
\value{
logical
}
\description{
Is this repo a package
}
\examples{
is_package_repo("maelle", "convertagd")
is_package_repo("maelle", "roregistry")
}
