% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spy.R
\name{spy}
\alias{spy}
\title{Get latest issues where an user commented}
\usage{
spy(user)
}
\arguments{
\item{user}{username}
}
\value{
tibble
}
\description{
Get latest issues where an user commented
}
\examples{
spy("github")
}
