% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contents.R
\name{get_contents}
\alias{get_contents}
\title{Get contents of a repo}
\usage{
get_contents(owner, repo)
}
\arguments{
\item{owner}{owner of the repo, string}

\item{repo}{repo name, string}
}
\description{
Get contents of a repo
}
\examples{
get_contents("maelle", "convertagd")
}
