% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_repos.R
\name{get_repos}
\alias{get_repos}
\title{Get repositories of one owner}
\usage{
get_repos(owner)
}
\arguments{
\item{owner}{string}
}
\value{
string of repos, format org/repo_name
}
\description{
Get repositories of one owner
}
\examples{
get_repos("jeroen")
}
