% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spy.R
\name{spy}
\alias{spy}
\title{Get latest issues where an user commented}
\usage{
spy(user, type = "Issue")
}
\arguments{
\item{user}{username}

\item{type}{"Issue" or "PullRequest"}
}
\value{
tibble
}
\description{
Get latest issues where an user commented
}
\examples{
# https://github.com/lintr-bot
spy("lintr-bot")
}
