% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spy.R
\name{spy}
\alias{spy}
\title{Get latest issues where an user commented}
\usage{
spy(user, type = "Issue", updated_after = as.character(Sys.Date() - 10),
  updated_before = as.character(Sys.Date()))
}
\arguments{
\item{user}{username}

\item{type}{"Issue" or "PullRequest"}

\item{updated_after}{date after which the issue or pull request should have
been updated(YYYY-MM-DD)}

\item{updated_before}{date before which the issue or pull request should have
been updated(YYYY-MM-DD)}
}
\value{
tibble
}
\description{
Get latest issues where an user commented
}
\examples{
# https://github.com/lintr-bot
spy("lintr-bot")
}
